Better handling of builtins
Builtins should be handled as special "bodies" of functions
so that we don't need to match function call names. In
addition to that, we should rely on code itself as definition
for builtins (without body) instead of hardcoding type of
builtin in type system. Then, it will be much easier to infer
identifier types corretcly and use them as functions in
folds/maps/filter parameters.