Fix type variable binding
Type variables (parameters) can be bound to each other. The reasoning for
this is to allow:
- Type parameters in function signatures that collapse to one type
- Type substitution
This in hindsight seems a bit pointless and dangerous, allowing many
mistypings and adding complexity by using unnecessary state.

Let's not bind type variables to other type variables. Doing so gets rid
of a lot of issues. Then later support for stateless substitution can be
added.

Also various rename to clarify code.