Fix argument expression type inference
This must be done in the caller context, otherwise inner context of
function shadows identifiers in the caller expressions. For example
if foldl defines a parameter x:List['T] and gets called in a function
f with x:OtherType as one argument, if we infer expression type in
the foldl context, we end up using x:List['T] instead of x:OtherType.