Substitute external type vars in result types
Currently, inner function expression type can resolve to a type
variable that is not part of the function signature. For example,
defining a function f that take a List['T] and calls a function g
that takes a List['A] and produces a List['A] resolves to a
return type of List['A] and has substitutions 'T -> 'A. This means
that all 'T become 'A which is correct, however when returning
the final type, we must invert the substitutions for all type vars
that appear in inputs and identifiers (external type vars), such
that 'A -> 'T.

As a follow up, an additional step of re-initialising all internal
type vars might be necessary, so that calling functions with
the same type vars 'T can resolve to different types.