Add simple error handling
A parser should fail in two ways: fail with retry, and fail with abort.
When a failure is not critical, i.e. there are other rules that can match
the input, the parser should fail with retry and continue with another
parser. This is the case for OR and for ambiguous rules.

However, for certain parsers it is obvious when there is no point
continuing parsing and we should abort with a clear abort message.
In general, after matching unique keywords, we can abort on failure.
This is the case for if, with, cast and fn statements, where if we matched
the keyword, there is no other rule possible.

We can then manually decorate these parsers with an aborting parser
which returns a clear abort message when failing parsing. This will
then propagate up to the caller.