Don't override bindings in type system
When a type can be bound to a type variable, we always rebind
the binding even though we should error out if a non-compatible
binding already exists. For example, we can currently bind Int to
'T and then String to 'T immediately after. This is because we are
not returning false in TypeParameterBindings::is_subtype when
the abstract binding and the concrete type are not compatible,
and instead we are returning true and inserting the new binding.
Let's fix this bug by returning the result of is_subtype.