Accept first arg as input file name
When laxma-rs is invoked with no arguments, it should open the REPL.
When laxma-rs is invoked with many arguments, the first one should be used as the
input file name to execute. In this case, the file is expected to contain a main function
which is the entry point.

Every laxma file can then be used as an executable script with shebang.