Improve error message with wrong destructuring
When mistakenly destructuring a function type, we get a "cannot find type"
error which is confusing as it points to the fact that the function type of an
argument is wrong, while instead the destructuring is wrong (for example
trying to match None pattern on a function type). Let's clarify the message.